// github/types/paths/repos/{owner}/{repo}.types.ts
export type reposGet = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: reposGet_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      200: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: full_repository;
          };
        };
        examples: {
          "default-response": unknown;
        };
      };
      301: moved_permanently;
      403: forbidden;
      404: not_found;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;
// ...
export type reposDelete = (
  $: OmitValueWhenNever<{
    query: never;
    querystring: never;
    path: reposDelete_Path;
    headers: never;
    cookie: never;
    body: never;
    context: Context;
    response: ResponseBuilderFactory<{
      204: {
        headers: never;
        requiredHeaders: never;
        content: never;
        examples: {};
      };
      307: temporary_redirect;
      403: {
        headers: never;
        requiredHeaders: never;
        content: {
          "application/json": {
            schema: { message?: string; documentation_url?: string };
          };
        };
        examples: {
          default: unknown;
        };
      };
      404: not_found;
      409: conflict;
    }>;
    x: WideOperationArgument;
    proxy: (url: string) => COUNTERFACT_RESPONSE;
    user: never;
    delay: (milliseconds: number, maxMilliseconds?: number) => Promise<void>;
    version: never;
  }>,
) => MaybePromise<COUNTERFACT_RESPONSE>;

export type reposGet_Path = { owner: string; repo: string };

export type reposUpdate_Path = { owner: string; repo: string };

export type reposDelete_Path = { owner: string; repo: string };